Move inpainting code to its own package
