+++ render-app/src/main/java/org/janelia/alignment/inpainting/RandomRayDirection2D.java
@@ -1,4 +1,4 @@
-package org.janelia.alignment.destreak;
+package org.janelia.alignment.inpainting;
 
 import java.util.Random;
